requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
DateTime==4.4
Flask==2.0.3
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
toml==0.10.2
urllib3==1.26.8
Werkzeug==2.0.3
zope.interface==5.4.0
